Task 1
TASK 1:- Multiplication Table Of Any Number.

Here are the simplified steps to perform multiplication of any number in Java:

Prompt the user to enter a number.
Read and store the number entered by the user.
Prompt the user to enter another number to multiply with the first number.
Read and store the second number entered by the user.
Perform the multiplication of the two numbers. 6. Display the result of the multiplication to the user.